backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
trimesh==4.1.2
numpy==1.26.4
cadquery-ocp==7.9.3.1
gmsh==4.12.1
requests==2.31.0
python-dotenv==1.0.1
